CAS interface service: remove razor pages, address build warnings, combine startup and program
