--- CASInterfaceService/CASInterfaceService/Pages/About.cshtml
@@ -1,9 +1,0 @@
-﻿@page
-@model AboutModel
-@{
-    ViewData["Title"] = "About";
-}
-<h2>@ViewData["Title"]</h2>
-<h3>@Model.Message</h3>
-
-<p>Use this area to provide additional information.</p>
