--- CASInterfaceService/CASInterfaceService/Pages/Index.cshtml.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace CASInterfaceService.Pages
-{
-    public class IndexModel : PageModel
-    {
-        public void OnGet()
-        {
-
-        }
-    }
-}
