--- CASInterfaceService/CASInterfaceService/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using CASInterfaceService
-@namespace CASInterfaceService.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
